requirements.txt
neo4j_arrow@https://github.com/neo4j-field/neo4j_arrow/archive/refs/tags/0.4.0.tar.gz
google-dataproc-templates>=0.0.1
google-cloud-bigquery-storage[pyarrow]>=2.18
graphdatascience==1.6
pyarrow>=4,<11
fsspec[gcs]>=2023.1.0
